# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Bandjacks API",
    "description": "\n    ## Cyber Threat Defense World Modeling API\n    \n    Bandjacks provides comprehensive cyber threat intelligence operations including:\n    \n    - **ATT&CK Data Management**: Load and manage MITRE ATT&CK releases\n    - **Natural Language Search**: Hybrid vector and graph search for CTI\n    - **Graph Exploration**: Traverse and analyze threat relationships\n    - **Document Extraction**: Extract CTI entities from documents using LLMs\n    - **Review Workflows**: Human-in-the-loop validation and feedback\n    - **Defense Recommendations**: D3FEND integration (coming in Sprint 4)\n    \n    ### Key Features\n    \n    - STIX 2.1 compliant data model\n    - ADM (ATT&CK Data Model) validation\n    - Hybrid search with vector embeddings and graph patterns\n    - LLM-powered extraction with evidence grounding\n    - Review queue for candidate validation\n    - Performance optimized with caching and connection pooling\n    \n    ### Authentication\n    \n    Currently no authentication required (development mode).\n    Production deployments should implement appropriate authentication.\n    \n    ### Rate Limiting\n    \n    No rate limiting currently implemented.\n    Production deployments should add rate limiting.\n    ",
    "version": "1.0.0",
    "contact": {
      "name": "Bandjacks Team",
      "email": "support@bandjacks.io"
    },
    "license": {
      "name": "Apache 2.0",
...
    }
...
  "servers": [
...
  "paths": {
...
    "/v1/extract/report": {
      "post": {
        "tags": [
          "extraction"
        ],
        "summary": "Extract Report",
        "description": "Extract CTI from a report with full provenance tracking.\n\nThis endpoint:\n1. Registers the source document with hash-based ID\n2. Chunks the text for processing\n3. Extracts entities and relationships using LLM\n4. Resolves entities to existing KB entries\n5. Builds STIX 2.1 bundle with provenance\n6. Optionally ingests to Neo4j graph",
        "operationId": "extract_report_v1_extract_report_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/extract/provenance/{object_id}": {
      "get": {
        "tags": [
          "extraction"
        ],
        "summary": "Get Object Provenance",
        "description": "Get complete provenance history for a STIX object.\n\nThis traces the object back through all extraction sources.",
        "operationId": "get_object_provenance_v1_extract_provenance__object_id__get",
        "parameters": [
          {
            "name": "object_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Object Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Object Provenance V1 Extract Provenance  Object Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ExtractionRequest": {
        "properties": {
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url",
            "description": "URL of source document"
          },
          "source_type": {
            "type": "string",
            "title": "Source Type",
            "description": "Type of source (report, blog, pdf, etc)",
            "default": "report"
          },
          "content": {
            "type": "string",
            "title": "Content",
            "description": "Text content to extract from"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title",
            "description": "Optional title for the report"
          },
          "method": {
            "type": "string",
            "title": "Method",
            "description": "Extraction method (llm, vector, hybrid)",
            "default": "llm"
          },
          "confidence_threshold": {
            "type": "number",
            "title": "Confidence Threshold",
            "description": "Minimum confidence for extraction",
            "default": 50.0
          },
          "auto_ingest": {
            "type": "boolean",
            "title": "Auto Ingest",
            "description": "Automatically ingest to graph if True",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "ExtractionRequest",
        "description": "Request for report extraction."
      },
      "ExtractionResponse": {
        "properties": {
          "extraction_id": {
            "type": "string",
            "title": "Extraction Id"
          },
          "source_id": {
            "type": "string",
            "title": "Source Id"
          },
          "bundle": {
            "additionalProperties": true,
            "type": "object",
            "title": "Bundle"
          },
          "stats": {
            "additionalProperties": true,
            "type": "object",
            "title": "Stats"
          },
          "provenance": {
            "additionalProperties": true,
            "type": "object",
            "title": "Provenance"
          },
          "ingested": {
            "type": "boolean",
            "title": "Ingested",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "extraction_id",
          "source_id",
          "bundle",
          "stats",
          "provenance"
        ],
        "title": "ExtractionResponse",
        "description": "Response from extraction."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT authentication (future implementation)"
      },
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "API key authentication (future implementation)"
      }
    },
...
  },
...
}